+++ Changelog.md
@@ -78,7 +78,7 @@
 - Hide Auto Layout annealing settings by default
 - Calculate mesh volume asynchronously (helps on models with ~ >10M faces)
 - Reduce unnecessary cross-thread memory copies
-- Update slicer CLI progress reporting
+- Update slicer CLI progress reporting ([#39](https://github.com/connorslade/mslicer/pull/39))
 
 ## v0.9.2 &mdash; August 30th, 2026